requirements.txt
asgiref==3.5.2
astroid==2.15.5
black==23.3.0
click==8.1.3
colorama==0.4.6
dill==0.3.6
Django==4.1.9
django-stubs==4.2.0
django-stubs-ext==4.2.0
django-taggit==3.0.0
isort==5.12.0
lazy-object-proxy==1.9.0
Markdown==3.4.1
mccabe==0.7.0
mypy==1.3.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
psycopg2-binary==2.9.3
pylint==2.17.4
pylint-django==2.5.3
pylint-plugin-utils==0.8.1
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.8
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.9
typing_extensions==4.5.0
tzdata==2023.3
wrapt==1.15.0
